FSM-#3 modify playLoop and implement fsm functions
