+++ inc/utils.h
@@ -37,4 +37,5 @@ inline string parseMicroSec(long us) {
     return string(buf);
 }
 
+
 #endif
